requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
django-cron==0.6.0
django-filter==23.3
django-quill-editor==0.1.42
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
psycopg2-binary
psycopg2